requirements.txt
django-cors-headers==3.13.0
    Django==4.1.4
        asgiref==3.5.2
        sqlparse==0.4.3
django-rest-framework==0.1.0
    djangorestframework==3.14.0
        # Django==4.1.4
            # asgiref==3.5.2
            # sqlparse==0.4.3
        pytz==2022.6
pipfreeze==2.0.3
pipreqs==0.4.11
    docopt==0.6.2
    yarg==0.1.9
        requests==2.28.1
            certifi==2022.12.7
            charset-normalizer==2.1.1
            idna==3.4
            urllib3==1.26.13
psycopg2==2.9.5
python-dateutil==2.8.2
    six==1.16.0
simplejson==3.18.0
